ai-ledger/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
asyncpg==0.29.0
pydantic[email]==2.10.0
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
anthropic==0.42.0
httpx==0.28.0
pyotp==2.9.0
reportlab==4.2.0
openpyxl==3.1.5
python-multipart==0.0.18
Pillow==11.1.0
stripe==11.4.0
python-dateutil==2.9.0
